fix:Effectクラスをパーティクル以外にも対応できるように抽象クラス化
